Resolved circular import issue
